+++ Editor/Interfaces/IBuildCache.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using UnityEditor.Build.Content;
 using UnityEditor.Build.Pipeline.Utilities;
+using UnityEngine;
 
 using UnityEngine;
 using UnityEditor;