+++ Editor/Interfaces/IBuildContent.cs
@@ -4,6 +4,7 @@ using UnityEngine;
 using UnityEditor;
 using UnityEditor.Build.Content;
 using UnityEditor.Build.Pipeline.Tasks;
+using UnityEngine;
 
 namespace UnityEditor.Build.Pipeline.Interfaces
 {