+++ Editor/Shared/BuildExtendedAssetData.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using UnityEditor.Build.Pipeline.Interfaces;
+using UnityEngine;
 
 using UnityEngine;
 using UnityEditor;