+++ Editor/Tasks/GenerateSubAssetPathMaps.cs
@@ -5,6 +5,7 @@ using UnityEditor.Build.Pipeline.Injector;
 using UnityEditor.Build.Pipeline.Interfaces;
 using UnityEditor.Build.Pipeline.Utilities;
 using UnityEditor.Build.Pipeline.WriteTypes;
+using UnityEngine;
 
 using UnityEngine;
 using UnityEditor;