+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [2.6.0] - 2026-02-16
+- Added support for extracting type tree data from asset bundles.  This requires Unity 6.5+.
+- Avoid pulling in exta dependencies when using recursive dependencies
+
 ## [2.5.2] - 2026-01-05
 - Upgrade references from UnityEditor.GUID to UnityEngine.GUID on Unity 6.4+.